Change the output of the CityList WS.
Before this webservice was returning data which could only by use for one
country. Now we have a country and a state field to differentiate.

Also modify all the code using this webservice data.

The old format is still available if you don't pass the new "version"
field or if you pass it with the value "1"