Fix the requestList onResume code.
The list size optimization was a mistake as the callListenerWithCachedData
can reduce the size of the list during the for loop as it is calling
the listener synchronously.
The removal of the request is also added in the onResume to make it more
clear that it is removed. Also if the developer forgets to remove it
in the listener, it won't be a problem as it is removed in the onResume.

The only downside is that the listener code is going to try to remove
the request which is going to be already removed. But it's not a big
problem as the requestList is going to be pretty small.