Add a sample using the Basic Authentication.
This sample uses the Basic HTTP Authentication available in
NetworkConnection. Check AuthenticationOperation to see how to use it.

Possible improvement : add a way to give back the 401 status code to the
onRequestConnectionError method so that the developer can detect that
he needs a Basic Authentication ? Not sure it's really useful as normally
the developer knows his webservice and will provide himself the UI to
the user.