Add a delay before stopping the Service
This patch ensures the MultiThreadedIntentService won't stop itself
as soon as no task need to be processed. By waiting for 30 seconds,
the system may reuse the Service once a new task is sent. As a
consequence, this patch may relieve the system from instantiating
new Services over and over again