Fix an issue with transfering custom Parcelable inside the Request
The ClassLoader used to read the Parcelable object from the Request object
was not the right one and therefore didn't know how to read the object if
you were using your own custom Parcelable. (If you are using SDK
Parcelable like Bitmap for example, it was already working)

This is now fixed and will work even for your custom Parcelable objects.

Fixes #14